backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langgraph
langchain-groq
python-dotenv
pydantic
